DirectFlow -> for filtering cli parameters + stripcmp -> strip and see if a particular char is in a string

+++ src/source/program_actions.c
@@ -1,11 +1,37 @@
 #include "../header/main.h"
 
+
+//based on the cli parameteres provided on runtime, act accordingly
+void DirectFlow(int argc, char *argv[])
+{
+    switch(argc)
+    {
+        case 1:
+            fprintf(stderr, "Error - no parameters provided\n");
+            break;
+        
+        case 2:
+            if(!(stripcmp(argv, '-')))
+            {
+                struct Logistics *core_logistics = InitLogistics();
+                DeleteFile(core_logistics, argv[1]); 
+                free(core_logistics);
+                break;  
+            }
+
+            fprintf(stderr, "%s missing argument", argv[1]);
+            break;
+
+        default:
+            ParseFlags(argc, argv);
+            break;
+    }
+}
+
 //flags dictate the next function call on the target file
 void ParseFlags(int argc, char *argv[])
 {
-    //sponge
-    // printf("%s", argv[1]);
-    printf("%d", argc);
+    struct Logistics *core_logistics = InitLogistics();
 
     char *use_flag = argv[1]; 
 
@@ -21,42 +47,44 @@ void ParseFlags(int argc, char *argv[])
         {"-V", "--VERBOSE_VIEW"}
     };
 
+    //the option to break helps limit the numbers of checks we have to make each flag
+    while (1) {
 
-    //sponge
-    printf("%c", flags[RESTORE][CONCISE][0]);
+        if (cmpstr(flags[RESTORE][CONCISE], use_flag) ||
+            cmpstr(flags[RESTORE][VERBOSE], use_flag) ){
+                printf("restore");                
+                break;        
+        }
 
-    struct Logistics *core_logistics = InitLogistics();
-
-    if (cmpstr(flags[RESTORE][CONCISE], use_flag) ||
-        cmpstr(flags[RESTORE][VERBOSE], use_flag) ){
-            printf("restore");
-    }
-    
-    if (cmpstr(flags[ORIGINAL_RESTORE][CONCISE], use_flag) ||
-        cmpstr(flags[ORIGINAL_RESTORE][VERBOSE], use_flag) ){
-            printf("original_restore");
-    }
+        if (cmpstr(flags[ORIGINAL_RESTORE][CONCISE], use_flag) ||
+            cmpstr(flags[ORIGINAL_RESTORE][VERBOSE], use_flag) ){
+                printf("original_restore");                
+                break;
+        }
         
-    if (cmpstr(flags[DELETE][CONCISE], use_flag) ||
-        cmpstr(flags[DELETE][VERBOSE], use_flag) ){
-            DeleteFile(InitLogistics(), argv[2]);
-            printf("delete");
-    }
-    
-    if (cmpstr(flags[VIEW_TRASH][CONCISE], use_flag) ||
-        cmpstr(flags[VIEW_TRASH][VERBOSE], use_flag) ){
-            ViewTrash(core_logistics, CONCISE);
-    }
+        if (cmpstr(flags[DELETE][CONCISE], use_flag) ||
+            cmpstr(flags[DELETE][VERBOSE], use_flag) ){
+                DeleteFile(InitLogistics(), argv[2]);                
+                break;        
+        }
+        
+        if (cmpstr(flags[VIEW_TRASH][CONCISE], use_flag) ||
+            cmpstr(flags[VIEW_TRASH][VERBOSE], use_flag) ){
+                ViewTrash(core_logistics, CONCISE);                
+                break;   
+        }
 
-    if (cmpstr(flags[VERBOSE_VIEW][CONCISE], use_flag) ||
-        cmpstr(flags[VERBOSE_VIEW][VERBOSE], use_flag) ){
-            ViewTrash(core_logistics, VERBOSE);
-    }
+        if (cmpstr(flags[VERBOSE_VIEW][CONCISE], use_flag) ||
+            cmpstr(flags[VERBOSE_VIEW][VERBOSE], use_flag) ){
+                ViewTrash(core_logistics, VERBOSE);
+                break;      
+        }
 
-    else 
-    {
-        fprintf(stderr, "Null use flags specified\n");
-    }   
+        //making it this far is an error on the users part
+        fprintf(stderr, "%s invalid flag", use_flag);
+        break;
+
+    }
 
     free(core_logistics);
 }
